Changed directory structure of mathx
