+++ mathx.h
@@ -1,7 +1,5 @@
 #include <cmath>
-#include "lib/number_theory.h"
-#include "lib/arithmetic.h"
-#include "lib/algebra.h"
-#include "lib/calculus.h"
-#include "lib/data_structures.h"
-#include "lib/int_l.h"
+#include <mathx/algebra.h>
+#include <mathx/int_l.h>
+#include <mathx/data_structures.h>
+#include <mathx/int_l.h>
