+++ src/number_theory.cpp
@@ -1,0 +1,19 @@
+#ifndef _NUMBER_THEORY_H_
+#define _NUMBER_THEORY_H_
+
+namespace mathx {
+    int gcd(int a, int b) {
+        while( a * b ){
+            if(a > b)
+                a = a % b;
+            else
+                b = b % a;
+        }
+        return a + b;
+    }
+    int lcm(int a, int b) {
+        return (a * b / gcd(a,b));
+    }
+}
+
+#endif /* _NUMBER_THEORY_H_ */
